Move services
Move the application services to the ‘providers’ folder
