+++ src/providers/order.service.spec.ts
@@ -3,10 +3,10 @@ import { Http, HttpModule, BaseRequestOptions,
   Response, ResponseOptions } from '@angular/http';
 import { MockBackend } from '@angular/http/testing';
 
-import { OrderState } from './commons';
-import { Order } from './order';
-import { OrderItem } from './order-item';
-import { Product } from './product';
+import { OrderState } from '../app/commons';
+import { Order } from '../app/order';
+import { OrderItem } from '../app/order-item';
+import { Product } from '../app/product';
 import { AuthMock, OrderMock } from '../test/mocks';
 
 import { OrderService } from './order.service';
